requirements.txt
numpy==1.26.4
streamlit==1.32.0
plotly==5.18.0
streamlit-lottie==0.0.5
langchain==0.1.12
spacy==3.7.4
pypdf2==3.0.1
python-docx==1.0.1
ollama==0.1.6
pydantic>=2.0.0
aiohttp==3.9.1
scikit-learn==1.4.1.post1
python-magic==0.4.27
tqdm==4.66.1
